split file
extension can now get access to realm
